# openapi.yml
...
openapi: 3.0.3
info:
  title: Uplink Connect API
  description: |
    Cloudflare-native data ingestion platform API.
    
    ## Authentication
    - **External endpoints**: Bearer token in `Authorization` header
    - **Internal endpoints**: `x-uplink-internal-key` header
  version: 3.01.0
  contact:
    name: BoringWorks
    url: https://github.com/Boring-Works/uplink-connect
servers:
  - url: https://uplink-core.codyboring.workers.dev
    description: Production Core API
  - url: https://uplink-edge.codyboring.workers.dev
    description: Production Edge API
...
paths:
...
  /internal/sources:
    get:
      tags: [Sources]
      summary: List sources
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: includeDeleted
          in: query
          schema: { type: boolean, default: false }
      responses:
        200:
          description: List of sources
    post:
      tags: [Sources]
      summary: Create or update source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceId, name, type]
              properties:
                sourceId: { type: string }
                name: { type: string }
                type: { type: string, enum: [api, webhook, email, file, browser, manual, stream] }
                config: { type: object }
      responses:
        201:
          description: Source created
  /internal/sources/{sourceId}:
    delete:
      tags: [Sources]
      summary: Delete source
      parameters:
        - name: sourceId
          in: path
          required: true
          schema: { type: string }
        - name: permanent
          in: query
          schema: { type: boolean, default: false }
      responses:
        200:
          description: Source deleted
  /internal/sources/{sourceId}/restore:
    post:
      tags: [Sources]
      summary: Restore deleted source
      parameters:
        - name: sourceId
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Source restored
  /internal/sources/{sourceId}/trigger:
    post:
      tags: [Sources]
      summary: Trigger collection workflow
      parameters:
        - name: sourceId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                force: { type: boolean }
                triggeredBy: { type: string }
                reason: { type: string }
      responses:
        202:
          description: Workflow started
...
  /internal/sources/{sourceId}/runs/tree:
    get:
      tags: [Sources]
      summary: Get source run tree
      parameters:
        - name: sourceId
          in: path
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        200:
          description: Run tree

  # Runs
  /internal/runs:
    get:
      tags: [Runs]
      summary: List runs
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        200:
          description: List of runs
  /internal/runs/{runId}:
    get:
      tags: [Runs]
      summary: Get run details
      parameters:
        - name: runId
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Run details
        404:
          description: Run not found
  /internal/runs/{runId}/replay:
    post:
      tags: [Runs]
      summary: Replay a run
      parameters:
        - name: runId
          in: path
          required: true
          schema: { type: string }
      responses:
        202:
          description: Replay queued
        404:
          description: Run not found
        409:
          description: Run cannot be replayed
  /internal/runs/{runId}/trace:
    get:
      tags: [Runs]
      summary: Get run trace
      parameters:
        - name: runId
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Run trace
        404:
          description: Run not found

  # Entities
...
  /internal/errors:
    get:
      tags: [Errors]
      summary: List ingest errors
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: sourceId
          in: query
          schema: { type: string }
        - name: phase
          in: query
          schema: { type: string }
        - name: errorCategory
          in: query
          schema: { type: string }
        - name: fromDate
          in: query
          schema: { type: string }
        - name: toDate
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        200:
          description: List of errors
  /internal/errors/{errorId}/retry:
    post:
      tags: [Errors]
      summary: Retry failed operation
      parameters:
        - name: errorId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                force: { type: boolean }
                triggeredBy: { type: string }
      responses:
        200:
          description: Retry succeeded
        409:
          description: Max retries exceeded
        422:
          description: Retry failed

  # Dashboard
...
  /internal/settings:
    get:
      tags: [Settings]
      summary: Get platform settings
      responses:
        200:
          description: Platform settings
    put:
      tags: [Settings]
      summary: Update platform settings
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        200:
          description: Settings updated
  /internal/audit-log:
    get:
      tags: [Settings]
      summary: Get audit log
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: resourceType
          in: query
          schema: { type: string }
        - name: actor
          in: query
          schema: { type: string }
        - name: fromDate
          in: query
          schema: { type: string }
        - name: toDate
          in: query
          schema: { type: string }
      responses:
        200:
          description: Audit log entries

  # Browser
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    InternalKey:
      type: apiKey
      in: header
      name: x-uplink-internal-key
security:
  - InternalKey: []